Homework 4 by Mardanov Amir
