+++ DynamicTable/Assets.xcassets/Companies logo/vkLogo.imageset/Contents.json
