+++ DynamicTable/Assets.xcassets/Post images/good.imageset/Contents.json
@@ -2,7 +2,7 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "arrow.png",
+      "filename" : "Kartinka_44_24043521.jpg",
       "scale" : "1x"
     },
     {
