--- DynamicTable/Constants.swift
@@ -1,17 +1,0 @@
-//
-//  Constants.swift
-//  DynamicTable
-//
-//  Created by Amir on 14.10.2019.
-//  Copyright © 2019 Ildar Zalyalov. All rights reserved.
-//
-
-import Foundation
-
-class Constants {
-    
-    static let showUserDetailInfo = "showUserDetailInfo"
-    static let showFullPost = "showFullPost"
-    static let editStatusSegue = "editStatus"
-    static let editPostSegue = "editPostSegue"
-}
