+++ DynamicTable/Models/Constants.swift
@@ -1,0 +1,19 @@
+//
+//  Constants.swift
+//  DynamicTable
+//
+//  Created by Amir on 14.10.2019.
+//  Copyright © 2019 Ildar Zalyalov. All rights reserved.
+//
+
+import Foundation
+
+class Constants {
+    
+    static let loginSegueId = "loginSegue"
+    static let showFullPostId = "showFullPost"
+    static let editStatusSegueId = "editStatus"
+    static let editPostSegueId = "editPostSegue"
+    static let showUserDetailSegueId = "showUserDetailInfo"
+    static let logoutSegueId = "logoutSegue"    
+}
